refactor(oauth-mcp): collapse 6-file package into single oauth_mcp.py
Mirrors supabase_tenant.py style — all logic in one self-contained module.
CF Access JWKS methods are private methods on OAuthMcpHttpExtension;
token helpers and metadata builders are module-level functions.
Public API and env var interface are unchanged.
